cli/mentatctl/requirements.txt
typer
PyYAML
jsonschema